fix: update tests after refactoring
